# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ONIS Solutions Code Challenge API using [newsapi.org](https://newsapi.org) Service.",
        "version": "1.0.0",
        "title": "ONIS-News-API",
        "contact": {
            "email": "dgarnica.dev@gmail.com"
        }
    },
    "host": "localhost:8080",
    "basePath": "/",
    "tags": [
        {
            "name": "News",
            "description": "Everything about recent news",
            "externalDocs": {
                "description": "Find out more",
                "url": "http://swagger.io"
            }
        }
    ],
    "schemes": ["http"],
    "paths": {
        "/api/news": {
            "get": {
                "tags": ["News"],
                "summary": "Find all recent news",
                "description": "Get all news",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Keyword to filter News",
                        "default": "",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Article"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/news/:title": {
            "get": {
                "tags": ["News"],
                "summary": "Finds News by title",
                "description": "Find specific news based on the title",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Keyword to filter News",
                        "default": "",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Article"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "Article": {
            "type": "object",
            "properties": {
                "source": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "string"
                        },
                        "name": {
                            "type": "string"
                        }
                    }
                },
                "author": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "urlToImage": {
                    "type": "string"
                }
            }
        }
    },
    "externalDocs": {
        "description": "Find out more about Swagger",
        "url": "http://swagger.io"
    }
}